Try physically based absorption + scattering

+++ src/main/resources/rs117/hd/utils/color_utils.glsl
@@ -217,5 +217,5 @@ vec3 packedHslToSrgb(int hsl) {
 }
 
 vec3 unpackSrgb(int srgb) {
-    return srgbToLinear(ivec3(srgb >> 16, srgb >> 8 & 0xFF, srgb & 0xFF) / 255.f);
+    return srgbToLinear(vec3(srgb >> 16, srgb >> 8 & 0xFF, srgb & 0xFF) / 255);
 }
